reformat grave code and make it more reasonable
