+++ README.md
@@ -12,3 +12,9 @@
 * 代码实现简洁，运行性能高得分高
 * 要求通过测试代码自我证明代码能够可靠运行并正确实现上述功能
 * 将整个目录生成相应的merkle树，并将该树的根打印出来，并存储到sqlite数据库里面（包括时间戳、目录路径）
+
+## 使用方法
+```
+go build
+./grave --help
+```
